+++ sigle/src/icons/index.tsx
@@ -1,103 +1,6 @@
-import { IconProps } from '@radix-ui/react-icons/dist/types';
+export * from './Radix';
 export * from './Heading1Light';
 export * from './Heading2Light';
 export * from './Heading3Light';
 export * from './ImageLight';
 export * from './QuoteLight';
-
-export const Heading1Icon = ({ width, height }: IconProps) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox="0 0 15 15"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M7.22192 2.48462C6.97339 2.48462 6.77192 2.68609 6.77192 2.93462C6.77192 3.18315 6.97339 3.38462 7.22192 3.38462H8.3505V7.29214H3.68127V3.38462H4.81807C5.0666 3.38462 5.26807 3.18315 5.26807 2.93462C5.26807 2.68609 5.0666 2.48462 4.81807 2.48462H1.45269C1.20416 2.48462 1.00269 2.68609 1.00269 2.93462C1.00269 3.18315 1.20416 3.38462 1.45269 3.38462H2.58127V12.1H1.45269C1.20416 12.1 1.00269 12.3015 1.00269 12.55C1.00269 12.7985 1.20416 13 1.45269 13H4.81807C5.0666 13 5.26807 12.7985 5.26807 12.55C5.26807 12.3015 5.0666 12.1 4.81807 12.1H3.68127V8.19214H8.3505V12.1H7.22192C6.97339 12.1 6.77192 12.3015 6.77192 12.55C6.77192 12.7985 6.97339 13 7.22192 13H10.5873C10.8358 13 11.0373 12.7985 11.0373 12.55C11.0373 12.3015 10.8358 12.1 10.5873 12.1H9.4505V3.38462H10.5873C10.8358 3.38462 11.0373 3.18315 11.0373 2.93462C11.0373 2.68609 10.8358 2.48462 10.5873 2.48462H7.22192Z"
-        fill="currentColor"
-      />
-      <path
-        d="M13.0213 12.512C13.0213 12.784 13.2453 13 13.5093 13C13.7813 13 13.9973 12.784 13.9973 12.512V7.77603C13.9973 7.50402 13.7893 7.28802 13.5093 7.28802H13.4933C13.3493 7.28802 13.2053 7.32002 13.0133 7.37603L12.1653 7.64803C11.9653 7.70403 11.8373 7.86403 11.8373 8.07203C11.8373 8.29603 12.0373 8.48803 12.2693 8.48803C12.3173 8.48803 12.3893 8.48003 12.4453 8.46403L13.0213 8.28802V12.512Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-};
-
-export const Heading2Icon = ({ width, height }: IconProps) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox="0 0 15 15"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M6.23388 2.48462C5.98535 2.48462 5.78388 2.68609 5.78388 2.93462C5.78388 3.18315 5.98535 3.38462 6.23388 3.38462H7.36246V7.29214H2.69323V3.38462H3.83003C4.07856 3.38462 4.28003 3.18315 4.28003 2.93462C4.28003 2.68609 4.07856 2.48462 3.83003 2.48462H0.464648C0.21612 2.48462 0.0146484 2.68609 0.0146484 2.93462C0.0146484 3.18315 0.21612 3.38462 0.464648 3.38462H1.59323V12.1H0.464648C0.21612 12.1 0.0146484 12.3015 0.0146484 12.55C0.0146484 12.7985 0.21612 13 0.464648 13H3.83003C4.07856 13 4.28003 12.7985 4.28003 12.55C4.28003 12.3015 4.07856 12.1 3.83003 12.1H2.69323V8.19214H7.36246V12.1H6.23388C5.98535 12.1 5.78388 12.3015 5.78388 12.55C5.78388 12.7985 5.98535 13 6.23388 13H9.59926C9.84779 13 10.0493 12.7985 10.0493 12.55C10.0493 12.3015 9.84779 12.1 9.59926 12.1H8.46246V3.38462H9.59926C9.84779 3.38462 10.0493 3.18315 10.0493 2.93462C10.0493 2.68609 9.84779 2.48462 9.59926 2.48462H6.23388Z"
-        fill="currentColor"
-      />
-      <path
-        d="M10.8494 12.536C10.8494 12.824 11.0734 13 11.3774 13H14.5534C14.7934 13 14.9854 12.808 14.9854 12.568C14.9854 12.328 14.7934 12.136 14.5534 12.136H12.2814L13.4334 11.176C14.4254 10.368 14.8894 9.88801 14.8894 9.00801V8.99201C14.8894 7.99201 14.1374 7.32001 13.0414 7.32001C12.1614 7.32001 11.6414 7.65601 11.1854 8.22401C11.1134 8.31201 11.0734 8.42401 11.0734 8.52801C11.0734 8.77601 11.2734 8.97601 11.5214 8.97601C11.6814 8.97601 11.7934 8.90401 11.8654 8.82401C12.2094 8.40801 12.5214 8.20001 12.9694 8.20001C13.4814 8.20001 13.8734 8.52001 13.8734 9.07201C13.8734 9.57601 13.5934 9.93601 12.8254 10.584L11.1054 12.048C10.9374 12.184 10.8494 12.344 10.8494 12.536Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-};
-
-export const Heading3Icon = ({ width, height }: IconProps) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox="0 0 15 15"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M6.29394 2.48462C6.04541 2.48462 5.84394 2.68609 5.84394 2.93462C5.84394 3.18315 6.04541 3.38462 6.29394 3.38462H7.42252V7.29214H2.75329V3.38462H3.89009C4.13862 3.38462 4.34009 3.18315 4.34009 2.93462C4.34009 2.68609 4.13862 2.48462 3.89009 2.48462H0.524707C0.276179 2.48462 0.074707 2.68609 0.074707 2.93462C0.074707 3.18315 0.276179 3.38462 0.524707 3.38462H1.65329V12.1H0.524707C0.276179 12.1 0.074707 12.3015 0.074707 12.55C0.074707 12.7985 0.276179 13 0.524707 13H3.89009C4.13862 13 4.34009 12.7985 4.34009 12.55C4.34009 12.3015 4.13862 12.1 3.89009 12.1H2.75329V8.19214H7.42252V12.1H6.29394C6.04541 12.1 5.84394 12.3015 5.84394 12.55C5.84394 12.7985 6.04541 13 6.29394 13H9.65932C9.90785 13 10.1093 12.7985 10.1093 12.55C10.1093 12.3015 9.90785 12.1 9.65932 12.1H8.52252V3.38462H9.65932C9.90785 3.38462 10.1093 3.18315 10.1093 2.93462C10.1093 2.68609 9.90785 2.48462 9.65932 2.48462H6.29394Z"
-        fill="currentColor"
-      />
-      <path
-        d="M12.9653 13C14.1333 13 14.9253 12.256 14.9253 11.224V11.208C14.9253 10.176 14.1253 9.74402 13.2773 9.64002L14.5973 8.27202C14.7653 8.10402 14.8773 7.96802 14.8773 7.73602C14.8773 7.47202 14.6693 7.30402 14.3893 7.30402H11.5253C11.2853 7.30402 11.0933 7.49602 11.0933 7.73602C11.0933 7.96802 11.2853 8.16002 11.5253 8.16002H13.5573L12.2453 9.56002C12.0933 9.71202 12.0373 9.83202 12.0373 9.96002C12.0373 10.192 12.2293 10.384 12.4613 10.384H12.6613C13.4453 10.384 13.9493 10.696 13.9493 11.248V11.264C13.9493 11.776 13.5333 12.112 12.9813 12.112C12.4613 12.112 12.0613 11.928 11.7093 11.576C11.6293 11.504 11.5253 11.448 11.3813 11.448C11.1253 11.448 10.9093 11.664 10.9093 11.92C10.9093 12.064 10.9813 12.2 11.0693 12.272C11.5253 12.712 12.1413 13 12.9653 13Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-};
-
-export const ListNumberedIcon = ({ width, height }: IconProps) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox="0 0 15 15"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M1.6012 5.29C1.6012 5.416 1.7032 5.518 1.8292 5.518C1.9552 5.518 2.0572 5.416 2.0572 5.29V3.598C2.0572 3.472 1.9552 3.37 1.8292 3.37H1.8232C1.7602 3.37 1.6942 3.385 1.6072 3.409L1.3102 3.496C1.2172 3.523 1.1572 3.601 1.1572 3.694C1.1572 3.802 1.2502 3.892 1.3582 3.892C1.3822 3.892 1.4152 3.886 1.4392 3.88L1.6012 3.838V5.29Z"
-        fill="currentColor"
-      />
-      <path d="M4.5 4.5H13.5" stroke="currentColor" strokeLinecap="round" />
-      <path
-        d="M0.89444 8.281C0.89444 8.419 0.99944 8.5 1.14344 8.5H2.28344C2.39444 8.5 2.48444 8.413 2.48444 8.302C2.48444 8.191 2.39444 8.101 2.28344 8.101H1.55144L1.89344 7.837C2.26544 7.549 2.44544 7.375 2.44544 7.024V7.018C2.44544 6.625 2.15744 6.37 1.72244 6.37C1.40444 6.37 1.20644 6.484 1.03844 6.679C1.00544 6.718 0.98444 6.772 0.98444 6.823C0.98444 6.94 1.07744 7.033 1.19444 7.033C1.26644 7.033 1.32044 7 1.35044 6.967C1.46744 6.841 1.56044 6.784 1.69244 6.784C1.85744 6.784 1.97444 6.88 1.97444 7.057C1.97444 7.231 1.87244 7.348 1.61144 7.561L1.01444 8.05C0.93644 8.113 0.89444 8.191 0.89444 8.281Z"
-        fill="currentColor"
-      />
-      <path d="M4.5 7.5H13.5" stroke="currentColor" strokeLinecap="round" />
-      <path
-        d="M1.70137 11.536C2.16937 11.536 2.45737 11.245 2.45737 10.849V10.843C2.45737 10.453 2.17837 10.303 1.89337 10.255L2.30437 9.86203C2.38837 9.78103 2.44237 9.71802 2.44237 9.60402C2.44237 9.47802 2.34337 9.40002 2.20837 9.40002H1.18537C1.07737 9.40002 0.987368 9.49002 0.987368 9.59802C0.987368 9.70602 1.07737 9.79602 1.18537 9.79602H1.83037L1.42237 10.207C1.35037 10.279 1.32337 10.333 1.32337 10.396C1.32337 10.504 1.41337 10.594 1.52137 10.594H1.59337C1.85137 10.594 2.00437 10.69 2.00437 10.861V10.867C2.00437 11.023 1.88137 11.122 1.70737 11.122C1.54237 11.122 1.41637 11.068 1.29637 10.966C1.26037 10.936 1.21537 10.909 1.14937 10.909C1.02937 10.909 0.927368 11.011 0.927368 11.131C0.927368 11.197 0.963368 11.26 1.00537 11.296C1.17637 11.443 1.40137 11.536 1.70137 11.536Z"
-        fill="currentColor"
-      />
-      <path d="M4.5 10.5H13.5" stroke="currentColor" strokeLinecap="round" />
-    </svg>
-  );
-};
